+++ server/src/main/java/com/fluffy/infra/persistence/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.fluffy.global.persistence;
+package com.fluffy.infra.persistence.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
