+++ server/src/main/java/com/fluffy/infra/redis/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.fluffy.global.redis.config;
+package com.fluffy.infra.redis.config;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.ObjectMapper;
