+++ server/src/main/java/com/fluffy/infra/storage/StorageClient.java
@@ -1,0 +1,8 @@
+package com.fluffy.infra.storage;
+
+public interface StorageClient {
+
+    String getPresignedUrl(String fileName);
+
+    void delete(String fileName);
+}
